requirements.txt
jupyter==1.0.0
matplotlib==1.5.1
networkx==1.11
PyYAML==3.11
requests==2.10.0
